+++ tests/testthat/helper-vctrs.R
@@ -56,3 +56,15 @@ expect_equal <- function(object, expected, ...,
 raw2 <- function(...) {
   as.raw(list_unchop(list2(...), ptype = integer()))
 }
+cpl2 <- function(...) {
+  # R 4.4.0 changed `as.complex(NA_real/integer/logical)` so that it always uses
+  # a `0` in the imaginary slot. While this is reasonable, it is annoying for
+  # comparison purposes in tests, where we typically propagate the `NA`. As of
+  # rlang 1.1.1, `cpl()` inherits this behavior change so we have a custom version
+  # here that works the same on all R versions.
+  # https://github.com/wch/r-source/commit/1a2aea9ac3c216fea718f33f712764afc34f6ee8
+  out <- list2(...)
+  out <- as.complex(out)
+  out[is.na(out)] <- complex(real = NA_real_, imaginary = NA_real_)
+  out
+}